ch(implement-code-review-feedback): implement feedback from code review -	rename files to match default export -	rename tests files to include .spec.js -	move utility function to a seperate folder -	change validation funcion to a class -	update validation message [finishes #157903393]
