--- server/tests/index.js
@@ -1,3 +1,0 @@
-import './UsersController';
-import './Authorization';
-import './RequestsController';
