+++ server/tests/index.spec.js
@@ -1,0 +1,3 @@
+import './usersController.spec';
+import './authorization.spec';
+import './requestsController.spec';
